requirements.txt
django==4.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==3.11.0
djongo==1.3.6
django-heroku==0.3.1
psycopg2==2.9.3
pymongo==3.12.3
dnspython==2.2.1